builder: update core dists and macOS build
Add `cmake` option `DISABLE_MACOS_PACKAGE_MANAGERS` to disable
automatically using macOS Homebrew, MacPorts or Fink.

Update version regexps on the SFML headers in `FindSFML.cmake` to allow
for a variable number of spaces between the macro and its definition.

Builder core:

Rename some of the directories and files to be a bit nicer.

Update all core dists possible to the latest release or commit when the
release is too old.

Set location of `perl` executable in the `$perl` variable and use it
wherever `perl` is invoked, also symlink the perl used to the build root
bin so that it can be found in `PATH`.

Also update the `perl` command used to install `App::cpanminus` for
newer versions `App::Cpan` that come with newer versions of brew `perl`.

Do the same for meson and ninja if installed as core deps.

Replace `zlib` with `zlib-ng` and add `--zlib-compat` to the
`CONFIGURE_OVERRIDE` for it.

Replace `pkgconfig` with `pkgconf`.

Use both pcre1 and pcre2.

Remove `libxml2-python` dist alias and rebuild `libxml2` `--with-python`
and `--with-python3` after `python` 2 and 3 have built.

Remove `libwavpack` as `ffmpeg` 5 no longer uses it as a dependency and
remove the flag to enable it from the `ffmpeg` `CONFIGURE_OVERRIDE`.

Add options to `PROJECT_ARGS` to disable `brew` etc. using the previously
mentioned new `cmake` option and to use the builder `wx-config` instead of
the `brew` one. Also remove the hack to unlink the `brew` wxWidgets before a
build and relink it after as the args fix the conflict.

Remove `-Dintrospection=false` from default `meson` args, because `meson`
throws an error for unknown options.

Add `CXX` and `LD` to default make args.

Unset `LIBRARY_PATH` before `make install`, because some `Makefile`s use
it as a path part erroneously.

Remove old `ccache` patch.

Move `ccache` and its deps `zstd` and `hiredis` to after `cmake`, which
is required to build it.

Build `cmake` without `ccache`, as it's now a dependency of `ccache`.

Add new configure override type `autoreconf_noargs`, which is the same
as `autoreconf` except that the configure args are not passed to
`autogen.sh`. Use this override for `pkgconf`.

Update expat `PRE_BUILD` hook for removing docs from build.

Add `PRE_BUILD` hook for glib to fetch the `gvdb` subproject.

Add `PRE_BUILD` hook for `hiredis`, a dependency of `ccache`, to build
`STATIC` instead of `SHARED` libraries via `cmake`.

Add `POST_BUILD` hook for `pkgconf` to symlink `bin/pkgconf` to
`bin/pkg-config` for compatibility.

Update `harfbuzz` `POST_BUILD` hook for `meson`.

Use `python(2|3) -m pip install <module>` instead of the `pip(2|3)`
script wrappers because for some reason I had problems with those,
perhaps because of conflicts with `brew` or `nix`.

Add `-lffi` to python3 `EXTRA_LDFLAGS`.

Add `--cc="$CC" -cxx="$CXX"` to the `ffmpeg` `CONFIGURE_OVERRIDE` because
it apparently does not read those from the environment.

Replace `docbook` 5.0 with 5.0.1 and add docbook 5.1.

Add `--with-library-bits=64` to `libicu` args.

macOS builder fixes (tested on High Sierra 10.13 and Catalina 10.15):

Add support for predefined BUILD_ROOT and append BUILD_ROOT_SUFFIX.

Allow pre-defined CHECKOUT directory for building the project sources.

Sanitize `PATH` before build to remove `/usr/local` etc., use
/usr/local/bin/brew explicitly for deps instead.

If llvm and clang is installed from nix or brew, prepend the path of the
highest version found to `PATH`.

Remove `-lm` from build environment as macOS has never had an `m` math
library and this causes problems in some config checks.

Add LLVM `libunwind` from the LLVM archive.

Add a patch for `glib`'s `meson.build` to statically link
`gettext`/`libintl`, adds `-liconv -framework CoreFoundation`.

Add an extra location to check for `stdint.h` in XCode.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>